Add files via upload
This is a uploaded html doc with the internal javascript. I have learned to use a few default functions as well as get most of the basic concepts that were intended to be learned by the Odin Project's assignment of the Rock Paper Scissors lesson. I have grown more familiarity with creating variables, making and declaring functions. I got to practice using for loops by spamming the console with Game Results at the end of the game instead of using a for loop for the game function of 5 round play as intended by the Odin Project. Instead of the game ending after 5 rounds using the for loop for the game function I used the reload function to refresh the page once a game has been done. I did this to make ends meet with the interactivity of the game. I will continue to use the youtubehelp code as a reference if I need to, which helped me learn a lot more.